requirements.txt
Django==2.1.3
django-configurations==2.1
dj-database-url==0.5.0
djangorestframework==3.9.0
django-filter==2.0.0

# Tests
django-nose==1.4.6
nose-progressive==1.5.1
